refactor: 모든 in-domain 유스케이스 tier-1 폴더로 이동
- usecsae + tiered architecture 구현
- usecase라고 이름을 짓든, service라고 이름을 짓든, facade라고 이름을 짓든, 어차피 계층은 무한히 생길 수 있음을 고려하여 tiered 방식으로 구현해보기로함.
