+++ src/coupon/application/use-cases/tier-1-in-domain/get-all-coupons.use-case.ts
