+++ src/coupon/application/use-cases/tier-1-in-domain/get-coupon-by-id.use-case.ts
