+++ src/coupon/application/use-cases/tier-1-in-domain/recover-user-coupon.use-case.spec.ts
