+++ src/coupon/application/use-cases/tier-1-in-domain/use-user-coupon.use-case.spec.ts
