+++ src/coupon/application/use-cases/tier-1-in-domain/validate-user-coupon.use-case.spec.ts
