+++ src/order/order.module.ts
@@ -35,14 +35,14 @@ import { OrderRecoveryService } from "./application/order-recovery.service";
 import { TransactionService } from "../common/services/transaction.service";
 
 // Domain Use Cases
-import { CreateOrderUseCase } from "./application/use-cases/create-order.use-case";
-import { ApplyDiscountUseCase } from "./application/use-cases/apply-discount.use-case";
-import { ChangeOrderStatusUseCase } from "./application/use-cases/change-order-status.use-case";
-import { GetOrderByIdUseCase } from "./application/use-cases/get-order-by-id.use-case";
-import { GetOrderByUserIdUseCase } from "./application/use-cases/get-order-by-user-id.use-case";
-import { FindStalePendingOrdersUseCase } from "./application/use-cases/find-stale-pending-orders.use-case";
-import { FindFailedOrdersUseCase } from "./application/use-cases/find-failed-orders.use-case";
-import { GetPopularProductsUseCase } from "./application/use-cases/get-popular-products.use-case";
+import { CreateOrderUseCase } from "./application/use-cases/tier-1-in-domain/create-order.use-case";
+import { ApplyDiscountUseCase } from "./application/use-cases/tier-1-in-domain/apply-discount.use-case";
+import { ChangeOrderStatusUseCase } from "./application/use-cases/tier-1-in-domain/change-order-status.use-case";
+import { GetOrderByIdUseCase } from "./application/use-cases/tier-1-in-domain/get-order-by-id.use-case";
+import { GetOrderByUserIdUseCase } from "./application/use-cases/tier-1-in-domain/get-order-by-user-id.use-case";
+import { FindStalePendingOrdersUseCase } from "./application/use-cases/tier-1-in-domain/find-stale-pending-orders.use-case";
+import { FindFailedOrdersUseCase } from "./application/use-cases/tier-1-in-domain/find-failed-orders.use-case";
+import { GetPopularProductsUseCase } from "./application/use-cases/tier-1-in-domain/get-popular-products.use-case";
 import { OrderStatApplicationService } from "./application/order-stat.service";
 
 @Module({
