+++ src/product/application/use-cases/tier-1-in-domain/confirm-stock.use-case.ts
