+++ src/product/application/use-cases/tier-1-in-domain/get-stock-reservations-by-key.use-case.ts
