+++ src/product/application/use-cases/tier-1-in-domain/reserve-stock.use-case.ts
