+++ src/user/user.module.ts
@@ -4,9 +4,9 @@ import { UserController } from "./infrastructure/http/user.controller";
 import { UserApplicationService } from "./application/services/user.service";
 import { UserRepository } from "./infrastructure/persistence/user.repository";
 import { UserTypeOrmEntity } from "./infrastructure/persistence/orm/user.typeorm.entity";
-import { GetUserByIdUseCase } from "./application/use-cases/get-user-by-id.use-case";
-import { GetUserByEmailUseCase } from "./application/use-cases/get-user-by-email.use-case";
-import { CreateUserUseCase } from "./application/use-cases/create-user.use-case";
+import { GetUserByIdUseCase } from "./application/use-cases/tier-1-in-domain/get-user-by-id.use-case";
+import { GetUserByEmailUseCase } from "./application/use-cases/tier-1-in-domain/get-user-by-email.use-case";
+import { CreateUserUseCase } from "./application/use-cases/tier-1-in-domain/create-user.use-case";
 import { AuthModule } from "@/auth/auth.module";
 import { WalletModule } from "@/wallet/wallet.module";
 import { TransactionService } from "@/common/services/transaction.service";
