+++ src/wallet/wallet.module.ts
@@ -5,12 +5,12 @@ import { TransactionService } from "../common/services/transaction.service";
 import { UserBalanceRepository } from "./infrastructure/persistence/use-balance.repository";
 import { PointTransactionRepository } from "./infrastructure/persistence/point-transaction.repository";
 import { WalletApplicationService } from "./application/wallet.service";
-import { ChargePointsUseCase } from "./application/use-cases/charge-points.use-case";
-import { RecoverPointsUseCase } from "./application/use-cases/recover-points.use-case";
-import { UsePointsUseCase } from "./application/use-cases/use-points.use-case";
-import { GetUserPointsUseCase } from "./application/use-cases/get-user-points.use-case";
-import { CreateUserBalanceUseCase } from "./application/use-cases/create-user-balance.use-case";
-import { ValidateUsePointsUseCase } from "./application/use-cases/validate-use-points.use-case";
+import { ChargePointsUseCase } from "./application/use-cases/tier-1-in-domain/charge-points.use-case";
+import { RecoverPointsUseCase } from "./application/use-cases/tier-1-in-domain/recover-points.use-case";
+import { UsePointsUseCase } from "./application/use-cases/tier-1-in-domain/use-points.use-case";
+import { GetUserPointsUseCase } from "./application/use-cases/tier-1-in-domain/get-user-points.use-case";
+import { CreateUserBalanceUseCase } from "./application/use-cases/tier-1-in-domain/create-user-balance.use-case";
+import { ValidateUsePointsUseCase } from "./application/use-cases/tier-1-in-domain/validate-use-points.use-case";
 import { UserBalanceTypeOrmEntity } from "./infrastructure/persistence/orm/user-balance.typeorm.entity";
 import { PointTransactionTypeOrmEntity } from "./infrastructure/persistence/orm/point-transaction.typeorm.entity";
 import { TypeOrmModule } from "@nestjs/typeorm";
